+++ obtieneMatriz.h
@@ -3,6 +3,6 @@
 
 int cuenta_columnas(char *puntero);
 int cuenta_filas(char *puntero, int columnas);
-void Construye_matriz(int filas, int columnas, char *puntero, int (*laberinto)[filas][columnas]);
+void Construye_matriz(int filas, int columnas, char *puntero, int short(*laberinto)[filas][columnas]);
 
 #endif
